Create new custom exceptions module
